--- a/Cargo.toml
+++ b/Cargo.toml
@@ -164,6 +164,9 @@ sqlparser = "0.49.0"
 sysinfo = "0.30.12"
 test-log = "0.2.16"
 tiktoken-rs = "0.5.9"
+derive_more = "1.0.0"
+heapless = "0.8.0"
+
 tokio = {version = "1.37.0", features = [
   "net",
   "time",