… Cargo.lock
… src/daft-csv/Cargo.toml
… src/daft-csv/src/compression.rs
… src/daft-csv/src/local/pool/fixed_capacity_vec.rs